test(e2e): run cli e2e tests in-process instead of one subprocess per test
- extract createProgram and a new runProgram(argv) into src/cli/program.ts so the e2e suite drives the CLI in-process; index.ts becomes a thin shebang entry and keeps GlobalOptions as a type-only export
- add tests/e2e/helpers/run-cli.ts, an in-process runner that captures process.stdout/stderr.write and console.* and restores every touched global in finally
- split the 2,349-line tests/e2e/cli.test.ts into six per-command-group in-process files sharing the helper, cutting tests/e2e/** file time from 126s to under 30s
- add tests/e2e/cli-subprocess-smoke.test.ts with five real-subprocess smokes (shebang, bundled bin, exit-code propagation, non-TTY color, mcp stdio startup)
- redirect createProgram imports in the cli and contract tests to program.js and pin program.ts import-purity in tests/unit/cli/program.test.ts
- memoize measureStartupModules and add a file-level timeout to startup-modules.test.ts so the guard stays fast under the new parallel profile
- sync cli and tests module READMEs and regenerate factual counts

Closes #225
