+++ prospec/ai-knowledge/module-map.yaml
@@ -214,7 +214,7 @@ modules:
       - 終端
       - entry point
     rationale: "Thin I/O layer: no business logic, delegates to services"
-    last_verified: 2026-08-29T03:56:02.423Z
+    last_verified: 2026-08-29T05:31:28.140Z
   - name: templates
     description: Handlebars template library — 17 skills + 7 shared partials, 28
       references, 1 agent-config, 5 change, 15 init/knowledge (74 `.hbs`
@@ -263,8 +263,8 @@ modules:
     rationale: Pure resources — no logic, consumed by lib/template.ts
     last_verified: 2026-08-28T17:03:36.190Z
   - name: tests
-    description: 4-layer test suite — 173 files, 4,358 tests (unit 3217 + contract
-      977 + integration 45 + e2e 119). Validates every module — format
+    description: 4-layer test suite — 181 files, 4,369 tests (unit 3219 + contract
+      977 + integration 45 + e2e 128). Validates every module — format
       contracts, the cli-first probe + station-command contracts, the drift
       engine, token corpus, and the MCP protocol over in-memory transport.
     paths:
@@ -300,4 +300,4 @@ modules:
       - 驗證
       - vitest
     rationale: Quality gate — validates all layers with pyramid coverage
-    last_verified: 2026-08-29T03:56:02.423Z
+    last_verified: 2026-08-29T05:31:28.140Z
