+++ prospec/index.md
@@ -28,7 +28,7 @@ Load these specific convention files only when their topics are relevant to the 
 | **services** | init, knowledge, change, archive, agent-sync, spec-sync, product, feature-map, triggers, language, measure, check, mcp, serve, status, change-log, change-status, change-scale, change-progress, review-merge, verify-record, learn, validate, spec-show, upgrade, cascade, auto-draft | 服務, 業務邏輯, business logic, execute pattern, use case, 量測報告, 漂移檢查, 真相層 | Active | Business logic — one `execute()` service per command — init / quickstart / upgrade, knowledge init + update, change story / plan / tasks / log / status / scale / progress, review merge, verify record, learn, validate, archive + spec-sync + finalize, agent-sync, measure, drift check (record-review / record-tests / escaped-defect / auto-draft modes), auto-draft, and the read-only MCP server. | Isolates business logic from I/O layer, enables testability | types, lib |
 | **cli** | commands, formatters, commander, output, preaction, measure, check, strict, mcp, stdio, archive, dry-run, station-commands, finalize, sanitize, auto-draft | 指令, 命令列, command line, 終端, entry point | Active | Thin CLI entry — 18 top-level Commander commands + 29 formatters that parse → call one service → format output; the cli-first station commands each add one command/formatter pair, never logic. | Thin I/O layer: no business logic, delegates to services | types, lib, services |
 | **templates** | handlebars, auto-draft, hbs, skills, agent-configs, recipe-first, loading-rules, references, change, stable-prefix, entry-gate, scale, kind, ci-workflow, flywheel, lessons-ledger, feature-map, category, grouping, cli-probe, cli-first | 模板, 範本, handlebars, template engine, resources, 穩定前綴, 知識同步閘門, 複雜度適配, CI 閘門 | Active | Handlebars template library — 17 skills + 7 shared partials, 28 references, 1 agent-config, 5 change, 15 init/knowledge (74 `.hbs` templates) — the source of every generated skill, README, and index; every skill delegates its deterministic steps to the CLI behind the shared `_cli-probe` partial. | Pure resources — no logic, consumed by lib/template.ts | — |
-| **tests** | vitest, memfs, unit, integration, contract, e2e, knowledge-format, skill-format, token-corpus, drift, lessons-harvest, mcp-server, in-memory-transport, cli-first, station-commands | 測試, 單元測試, test suite, 驗證, vitest | Active | 4-layer test suite — 173 files, 4,358 tests (unit 3217 + contract 977 + integration 45 + e2e 119). Validates every module — format contracts, the cli-first probe + station-command contracts, the drift engine, token corpus, and the MCP protocol over in-memory transport. | Quality gate — validates all layers with pyramid coverage | types, lib, services, cli, templates |
+| **tests** | vitest, memfs, unit, integration, contract, e2e, knowledge-format, skill-format, token-corpus, drift, lessons-harvest, mcp-server, in-memory-transport, cli-first, station-commands | 測試, 單元測試, test suite, 驗證, vitest | Active | 4-layer test suite — 181 files, 4,369 tests (unit 3219 + contract 977 + integration 45 + e2e 128). Validates every module — format contracts, the cli-first probe + station-command contracts, the drift engine, token corpus, and the MCP protocol over in-memory transport. | Quality gate — validates all layers with pyramid coverage | types, lib, services, cli, templates |
 
 _Table format: Module | Keywords | Aliases | Status | Description | Rationale | Depends On_
