+++ src/cli/program.ts
@@ -1,0 +1,155 @@
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { Command } from 'commander';
+import pc from 'picocolors';
+import { handleError } from './formatters/error-output.js';
+import { ConfigNotFound } from '../types/errors.js';
+import { registerInitCommand } from './commands/init.js';
+import { registerQuickstartCommand } from './commands/quickstart.js';
+import { registerUpgradeCommand } from './commands/upgrade.js';
+import { registerPrintTemplateCommand } from './commands/print-template.js';
+import { registerKnowledgeCommand } from './commands/knowledge.js';
+import { registerKnowledgeInitCommand } from './commands/knowledge-init.js';
+import { registerKnowledgeUpdateCommand } from './commands/knowledge-update.js';
+import { registerKnowledgeVerifyCommand } from './commands/knowledge-verify.js';
+import { registerAgentCommand } from './commands/agent-sync.js';
+import { registerConfigCommand } from './commands/config.js';
+import { registerChangeCommand } from './commands/change-story.js';
+import { registerChangePlanCommand } from './commands/change-plan.js';
+import { registerChangeTasksCommand } from './commands/change-tasks.js';
+import { registerChangeLogCommand } from './commands/change-log.js';
+import { registerChangeStatusCommand } from './commands/change-status.js';
+import { registerChangeScaleCommand } from './commands/change-scale.js';
+import { registerChangeProgressCommand } from './commands/change-progress.js';
+import { registerChangeAutoDraftCommand } from './commands/change-auto-draft.js';
+import { registerStatusCommand } from './commands/status.js';
+import { registerSpecCommand } from './commands/spec-show.js';
+import { registerArchiveCommand } from './commands/archive.js';
+import { registerReviewCommand } from './commands/review-merge.js';
+import { registerVerifyCommand } from './commands/verify-record.js';
+import { registerLearnCommand } from './commands/learn.js';
+import { registerValidateCommand } from './commands/validate.js';
+import { registerMeasureCommand } from './commands/measure.js';
+import { registerCheckCommand } from './commands/check.js';
+import { registerMcpCommand } from './commands/mcp.js';
+import { PROSPEC_VERSION } from '../types/version.js';
+import type { GlobalOptions } from './index.js';
+
+/**
+ * Commands that do NOT require .prospec.yaml to exist.
+ */
+const INIT_COMMANDS = new Set(['init', 'quickstart', 'help', 'print-template', 'config']);
+
+export function createProgram(): Command {
+  const program = new Command();
+
+  program
+    .name('prospec')
+    .description('Progressive Spec-Driven Development CLI')
+    .version(PROSPEC_VERSION)
+    .option('--verbose', 'Enable verbose output')
+    .option('-q, --quiet', 'Quiet mode (results only, suitable for CI/CD)')
+    .configureOutput({
+      outputError: (str, write) => write(pc.red(str)),
+    })
+    .showSuggestionAfterError()
+    .showHelpAfterError('(use --help for additional information)')
+    .exitOverride();
+
+  // preAction hook: check .prospec.yaml existence for non-init commands
+  program.hook('preAction', (_thisCommand, actionCommand) => {
+    const cmdName = actionCommand.name();
+    // Walk up to find if any ancestor is in INIT_COMMANDS
+    let cmd: Command | null = actionCommand;
+    while (cmd) {
+      if (INIT_COMMANDS.has(cmd.name())) return;
+      cmd = cmd.parent;
+    }
+
+    // Skip check for root program (e.g. --version, --help)
+    if (cmdName === 'prospec') return;
+
+    // `mcp serve --cwd <dir>` targets another project root, so the guard must
+    // resolve .prospec.yaml against that dir — not the directory the agent was
+    // launched from (which may not be a Prospec project at all).
+    const targetCwd = (actionCommand.opts() as { cwd?: string }).cwd;
+    if (targetCwd !== undefined) {
+      const configPath = path.join(targetCwd, '.prospec.yaml');
+      if (!fs.existsSync(configPath)) throw new ConfigNotFound(configPath);
+      return;
+    }
+
+    if (!fs.existsSync('.prospec.yaml')) {
+      throw new ConfigNotFound();
+    }
+  });
+
+  // Register subcommand groups
+  registerInitCommand(program);
+  registerQuickstartCommand(program);
+  registerUpgradeCommand(program);
+  registerPrintTemplateCommand(program);
+  const knowledge = registerKnowledgeCommand(program);
+  registerKnowledgeInitCommand(knowledge, program);
+  registerKnowledgeUpdateCommand(knowledge, program);
+  registerKnowledgeVerifyCommand(knowledge, program);
+  registerAgentCommand(program);
+  registerConfigCommand(program);
+  registerChangeCommand(program);
+  registerChangePlanCommand(program);
+  registerChangeTasksCommand(program);
+  registerChangeLogCommand(program);
+  registerChangeStatusCommand(program);
+  registerChangeScaleCommand(program);
+  registerChangeProgressCommand(program);
+  registerChangeAutoDraftCommand(program);
+  registerStatusCommand(program);
+  registerSpecCommand(program);
+  registerArchiveCommand(program);
+  registerReviewCommand(program);
+  registerVerifyCommand(program);
+  registerLearnCommand(program);
+  registerValidateCommand(program);
+  registerMeasureCommand(program);
+  registerCheckCommand(program);
+  registerMcpCommand(program);
+
+  return program;
+}
+
+/**
+ * Build the program, parse argv, and dispatch errors. Exit status is carried on
+ * `process.exitCode` (set by command actions and by `handleError`), NOT by
+ * calling `process.exit()` — so this runs safely in-process (the e2e suite
+ * drives it directly) as well as from the `index.ts` entry point.
+ */
+export async function runProgram(argv: string[]): Promise<void> {
+  const program = createProgram();
+
+  try {
+    await program.parseAsync(argv);
+  } catch (err) {
+    // Commander throws on --help/--version with exitOverride; ignore clean exits
+    if (
+      err instanceof Error &&
+      'exitCode' in err &&
+      (err as { exitCode: number }).exitCode === 0
+    ) {
+      return;
+    }
+
+    // Commander parse errors (unknown option, missing arg) already outputted
+    if (
+      err instanceof Error &&
+      'code' in err &&
+      typeof (err as { code: string }).code === 'string' &&
+      (err as { code: string }).code.startsWith('commander.')
+    ) {
+      process.exitCode = 1;
+      return;
+    }
+
+    const opts = program.opts<GlobalOptions>();
+    handleError(err, opts.verbose ?? false);
+  }
+}
